// cypress/e2e/api/Endpoints.ts
export class Endpoints {
    BASE_URL: string = 'https://demoqa.com/';
    ACCOUNT: string = 'Account/v1/';
    USER: string = 'User';
    GENERATE_TOKEN: string = 'GenerateToken';
    BOOKSTORE: string = 'BookStore/v1/';
    BOOKS: string = 'Books';
    BOOK: string = 'Book';
}

export const endpoints = new Endpoints();